Move tabs to upper directory
Too many tabs/tabs around for a single file...
