+++ core/utils/poly-look/src/react-components/index.js
@@ -1,5 +1,5 @@
 import ChipGroup from "./chipGroup/chipGroup.jsx";
-import { Tab, Tabs } from "./tabs/tabs.jsx";
+import { Tab, Tabs } from "./tabs.jsx";
 import BlockLegend from "./legends/blockLegend/blockLegend.jsx";
 import LineLegend from "./legends/lineLegend/lineLegend.jsx";
